requirements.txt
google-genai>=1.15.0
youtube-transcript-api>=1.2.4
yt-dlp>=2025.1.26
oss2>=2.19.1
python-dotenv>=1.0.1
Pillow>=10.4.0
streamlit>=1.41.1
